requirements.txt
requests
beautifulsoup4
selenium
webdriver-manager
yt-dlp
python-dateutil